Move all ingredients to separate folder (#93)
* minor changes to peak-smoothing algo

* move all ingredients to separate folder

* further simplification of import
