+++ src/snapred/backend/dao/calibration/CalibrationRecord.py
@@ -2,9 +2,9 @@ from typing import List, Optional
 
 from pydantic import BaseModel
 
-from snapred.backend.dao import ReductionIngredients
 from snapred.backend.dao.calibration.Calibration import Calibration
 from snapred.backend.dao.calibration.FocusGroupMetric import FocusGroupMetric
+from snapred.backend.dao.ingredients import ReductionIngredients
 from snapred.backend.dao.state import FocusGroupParameters
