+++ src/snapred/backend/dao/ingredients/PixelGroupingIngredients.py
