+++ src/snapred/backend/recipe/FitCalibrationWorkspaceRecipe.py
@@ -1,9 +1,11 @@
 import json
 from typing import Any, Dict
 
-from snapred.backend.dao.FitMultiplePeaksIngredients import FitMultiplePeaksIngredients
-from snapred.backend.dao.ingredients.FitCalibrationWorkspaceIngredients import FitCalibrationWorkspaceIngredients
-from snapred.backend.dao.SmoothDataExcludingPeaksIngredients import SmoothDataExcludingPeaksIngredients
+from snapred.backend.dao.ingredients import (
+    FitCalibrationWorkspaceIngredients,
+    FitMultiplePeaksIngredients,
+    SmoothDataExcludingPeaksIngredients,
+)
 from snapred.backend.log.logger import snapredLogger
 from snapred.backend.recipe.GenericRecipe import (
     FitMultiplePeaksRecipe,
