+++ src/snapred/backend/recipe/algorithm/AlignAndFocusReductionAlgorithm.py
@@ -1,7 +1,7 @@
 from mantid.api import AlgorithmFactory, PythonAlgorithm
 from mantid.kernel import Direction
 
-from snapred.backend.dao.ReductionIngredients import ReductionIngredients
+from snapred.backend.dao.ingredients import ReductionIngredients
 from snapred.backend.recipe.algorithm.MantidSnapper import MantidSnapper
 
 name = "AlignAndFocusReductionAlgorithm"
