+++ src/snapred/backend/recipe/algorithm/VanadiumFocussedReductionAlgorithm.py
@@ -3,8 +3,10 @@ import json
 from mantid.api import AlgorithmFactory, PythonAlgorithm, mtd
 from mantid.kernel import Direction
 
-from snapred.backend.dao.ReductionIngredients import ReductionIngredients
-from snapred.backend.dao.SmoothDataExcludingPeaksIngredients import SmoothDataExcludingPeaksIngredients
+from snapred.backend.dao.ingredients import (
+    ReductionIngredients,
+    SmoothDataExcludingPeaksIngredients,
+)
 from snapred.backend.recipe.algorithm.CustomGroupWorkspace import CustomGroupWorkspace  # noqa F401
 from snapred.backend.recipe.algorithm.MantidSnapper import MantidSnapper
 from snapred.backend.recipe.algorithm.SmoothDataExcludingPeaksAlgo import SmoothDataExcludingPeaks  # noqa F401
