+++ tests/unit/backend/recipe/algorithm/test_VanadiumFocussedReductionAlgorithm.py
@@ -11,8 +11,10 @@ with mock.patch.dict(
 ):
     from snapred.backend.dao.calibration.Calibration import Calibration
     from snapred.backend.dao.CrystallographicInfo import CrystallographicInfo
-    from snapred.backend.dao.ReductionIngredients import ReductionIngredients
-    from snapred.backend.dao.SmoothDataExcludingPeaksIngredients import SmoothDataExcludingPeaksIngredients
+    from snapred.backend.dao.ingredients import (
+        ReductionIngredients,
+        SmoothDataExcludingPeaksIngredients,
+    )
     from snapred.backend.recipe.algorithm.VanadiumFocussedReductionAlgorithm import (
         VanadiumFocussedReductionAlgorithm,  # noqa: E402
     )
