+++ tests/unit/backend/recipe/test_FitCalibrationWorkspaceRecipe.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import MagicMock, patch
 
-from snapred.backend.dao.ingredients.FitCalibrationWorkspaceIngredients import FitCalibrationWorkspaceIngredients
+from snapred.backend.dao.ingredients import FitCalibrationWorkspaceIngredients
 from snapred.backend.recipe.FitCalibrationWorkspaceRecipe import FitCalibrationWorkspaceRecipe
