+++ tests/unit/backend/recipe/test_SmoothDataExcludingPeaksRecipe.py
@@ -2,7 +2,7 @@ import unittest
 from unittest import mock
 
 import pytest
-from snapred.backend.dao.SmoothDataExcludingPeaksIngredients import SmoothDataExcludingPeaksIngredients
+from snapred.backend.dao.ingredients import SmoothDataExcludingPeaksIngredients
 from snapred.backend.recipe.SmoothDataExcludingPeaksRecipe import SmoothDataExcludingPeaksRecipe
